Colors

+++ controllers/users.js
@@ -2,6 +2,8 @@ const bcrypt = require('bcryptjs')
 const User = require('../models/User')
 const errorHandler = require('../utils/errorHandler')
 const { query } = require('express')
+const Institution = require('../models/Institution')
+const Message = require('../models/Message')
 
 module.exports.create = async function(req, res) {
     // login password
@@ -14,6 +16,10 @@ module.exports.create = async function(req, res) {
       })
     } else {
       // Нужно создать пользователя
+      let institution
+      if (req.user.levelStatus == 2) institution = req.user.institution
+      else institution = req.body.institution
+
       const salt = bcrypt.genSaltSync(10)
       const password = req.body.password
       const user = new User({
@@ -23,9 +29,22 @@ module.exports.create = async function(req, res) {
         surname: req.body.surname,
         birthDate: req.body.birthDate,
         sex: req.body.sex,
-        institution: req.body.institution,
+        institution: institution,
         levelStatus: req.body.levelStatus,
-        photo: req.file ? req.file.path : ''
+        photo: req.file ? req.file.path : (req.body.sex == '2' ? '/uploads/images/girl.png' : '/uploads/images/boy.png'),
+        online: req.body.online,
+        text: req.body.text,
+        read: req.body.read,
+        firstColor: req.body.firstColor,
+        secondColor: req.body.secondColor,
+        surnameView: req.body.surnameView,
+        setting: req.body.setting,
+        vote: req.body.vote,
+        sentence: req.body.sentence,
+        answers: req.body.answers,
+        change: req.body.change,
+        defaultColor: req.body.defaultColor,
+        birthdays: req.body.birthdays
       })
   
       try {
@@ -40,25 +59,9 @@ module.exports.create = async function(req, res) {
 
 module.exports.update = async function(req, res) {
   try {
-    if (req.body.login) {
-      const candidate = await User.findOne({login: req.body.login})
-  
-      if (candidate) {
-        // Пользователь существует, нужно отправить ошибку
-        res.status(409).json({
-          message: 'Такой логин уже занят. Попробуйте другой.'
-        })
-      }
-    } else {
-      const updated = {
-        login: req.body.login,
-        name: req.body.name,
-        surname: req.body.surname,
-        birthDate: req.body.birthDate,
-        sex: req.body.sex,
-        institution: req.body.institution,
-        levelStatus: req.body.levelStatus,
-      }
+    if (!req.body.login) {
+      const updated = req.body
+
       if (req.body.password) {
         const salt = bcrypt.genSaltSync(10)
         const password = req.body.password
@@ -76,16 +79,50 @@ module.exports.update = async function(req, res) {
       )
       res.status(200).json(thisuser)
     }
-  } catch (e) {
+    else {
+      const candidate = await User.findOne({login: req.body.login})
+  
+      if (candidate) {
+        // Пользователь существует, нужно отправить ошибку
+        res.status(409).json({
+          message: 'Такой логин уже занят. Попробуйте другой.'
+        })
+      } 
+      else {
+        const updated = req.body
+
+        if (req.body.password) {
+          const salt = bcrypt.genSaltSync(10)
+          const password = req.body.password
+          updated.password = bcrypt.hashSync(password, salt)
+        }
+
+        if (req.file) {
+          updated.photo = req.file.path
+        }
+        
+        const thisuser = await User.findOneAndUpdate(
+          {_id: req.params.userID},
+          {$set: updated},
+          {new: true}
+        )
+        res.status(200).json(thisuser)
+      }
+    }
+  }
+  catch (e) {
     errorHandler(res, e)
   }
 }
 
 module.exports.getAll = async function(req, res) {
   q = {}
-  
-  if (req.query.institution) {
-    q.institution =  req.query.institution
+  if (req.user.levelStatus != 1) {
+    q.institution = req.user.institution
+  } else {
+    if (req.query.institution) {
+      q.institution =  req.query.institution
+    }
   }
 
   if (req.query.name) {
@@ -101,47 +138,61 @@ module.exports.getAll = async function(req, res) {
   } 
 
   if (req.query.levelStatus) {
-    q.levelStatus =  req.query.levelStatus
+    let roles = req.query.levelStatus.split(',')
+    q.levelStatus = {$in: roles}
   } 
 
-  try {
-    const users = await User
-      .find(q)
-      .sort({surname: 1})
-    res.status(200).json(users)
-  } catch (e) {
-    errorHandler(res, e)
+  if (req.query.login) {
+    q.login =  req.query.login
   }
-}
 
-/*
-module.exports.getByInstitution = async function(req, res) {
   try {
-    const users = await User
-    .find({institution: req.params.institutionID})
-    .sort({surname: 1})
+    let users = await User
+      .find(q)
+      .sort({surname: 1, name: 1})
+      .skip(+req.query.offset)
+      .limit(+req.query.limit)
+      .lean()
+
+    for (let user of users) {
+      let institutionName = await Institution.findOne({_id: user.institution}, {_id: 0}) // {name: 'Name'}
+      user.institutionName = institutionName.name
+    }
+
     res.status(200).json(users)
   } catch (e) {
     errorHandler(res, e)
   }
 }
-*/
 
 module.exports.getByUserID = async function(req, res) {
   try {
-    const thisuser = await User.findOne({_id: req.params.userID})
-    res.status(200).json(thisuser)
+    let user = await User.findOne({_id: req.params.userID})
+    const institution = await Institution.findOne({_id: user.institution}, {_id: 0})
+    user.institutionName = institution.name
+    res.status(200).json(user)
   } catch (e) {
     errorHandler(res, e)
   }
 }
 
 module.exports.remove = async function(req, res) {
-  try {
-    await User.deleteOne({_id: req.params.userID})
-    res.status(200).json({
-      message: 'Пользоатель удален.'
-    })
+  try { 
+    if (req.params.userID != req.user.id) {
+      await User.deleteOne({_id: req.params.userID})
+      await Message.deleteMany({
+        $or: [{recipient: req.params.userID}, {sender: req.params.userID}]
+      })
+      res.status(200).json({
+        message: 'Пользователь удален.'
+      })
+    }
+    else {
+      res.status(403).json({
+        message: 'Вы не можете удалить себя.'
+      })
+    }
+    
   } catch (e) {
     errorHandler(res, e)
   }
